+++ client_test.go
@@ -316,8 +316,8 @@ func TestNewClient_UserAgent(t *testing.T) {
 			name: "with-service-account",
 			args: []any{
 				&endpoint.Authentication{
-					TenantID:       "fake-tenant-id",
-					ExternalIdPJWT: "fake-external-id-pjwt",
+					ExternalJWT: "fake-external-idp-jwt",
+					TokenURL:    "https://fake.token.url.com/token",
 				},
 			},
 		},