+++ pkg/endpoint/authentication.go
@@ -19,24 +19,24 @@ package endpoint
 // Authentication provides a struct for authentication data. Either specify User and Password for Trust Protection Platform
 // or Firefly or ClientId and ClientSecret for Firefly or specify an APIKey for TLS Protect Cloud.
 type Authentication struct {
-	//TPP Auth methods
-	// User and password
+	// TPP Auth methods
+	// user and password
 	User     string `yaml:"user,omitempty"`     //**DEPRECATED** Use access/refresh token or client certificate instead
 	Password string `yaml:"password,omitempty"` //**DEPRECATED** Use access/refresh token or client certificate instead
-	// Tokens
+	// tokens
 	AccessToken  string `yaml:"accessToken,omitempty"`
 	RefreshToken string `yaml:"refreshToken,omitempty"`
-	// Client certificate
+	// client certificate
 	ClientPKCS12 bool `yaml:"-"`
 
-	//TLSPC Auth methods
+	// VCP Auth methods
 	// API key
 	APIKey string `yaml:"apiKey,omitempty"`
 	// Service account
-	TenantID       string `yaml:"tenantId,omitempty"`
-	ExternalIdPJWT string `yaml:"externalJWT,omitempty"`
+	TokenURL    string `yaml:"tokenURL,omitempty"`
+	ExternalJWT string `yaml:"externalJWT,omitempty"`
 
-	// IDP Auth method
+	// OIDC Auth methods
 	ClientId     string `yaml:"clientId,omitempty"`
 	ClientSecret string `yaml:"clientSecret,omitempty"`
 	Scope        string `yaml:"scope,omitempty"`
@@ -46,6 +46,7 @@ type Authentication struct {
 
 // OAuthProvider provides a struct for the OAuth 2.0 providers information
 type OAuthProvider struct {
+	// OIDC Auth methods
 	DeviceURL string `yaml:"-"`
 	TokenURL  string `yaml:"tokenURL,omitempty"`
 	Audience  string `yaml:"audience,omitempty"`